requirements.txt
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
paho-mqtt>=1.6.0
streamlit>=1.28.0
